services/integration-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-jose
bcrypt==4.0.1
python-dotenv
pydantic[email]
redis
httpx